[Connection-Allowlist] Match simple patterns rather than origins.
This CL shifts the ductwork between //content and //services/network
to pass allowlist patterns as strings that are converted into
`SimpleUrlPatternMatcher` objects when injected into a given network
context for enforcement via `NetworkContext::RevokeNetworkForNonces`.
This allows us to support substantially more robust matching syntax, and
gets fairly close to the breadth of what URL Pattern promises.

In the future, we'll have to consider the fact that this implementation
via `SimpleUrlPatternMatcher` does not support regex. That might be a
problem as we move past the prototype phase, but it might also be a
reasonable tradeoff to make (and one in-line with shared dictionaries
and service worker routing). A question for later...

Bug: 447954811
Change-Id: Ibcc076ffb4c50a8d9b3c1a11fa08374abe83e844